backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
pymupdf
langchain-core
langchain-google-genai
python-multipart
google-generativeai